requirements.txt
streamlit
tensorflow
numpy
Pillow
gdown

